// powerapps-project-planner-20260612/src/generated/models/Vibe_tasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_tasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_tasksstatecode = keyof typeof Vibe_tasksstatecode;
export const Vibe_tasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_tasksstatuscode = keyof typeof Vibe_tasksstatuscode;
export const Vibe_tasksvibe_priority = {
  100000000: 'LOW',
  100000001: 'MEDIUM',
  100000002: 'HIGH',
  100000003: 'CRITICAL'
} as const;
export type Vibe_tasksvibe_priority = keyof typeof Vibe_tasksvibe_priority;
export const Vibe_tasksvibe_taskstatus = {
  100000000: 'TODO',
  100000001: 'IN_PROGRESS',
  100000002: 'BLOCKED',
  100000003: 'DONE'
} as const;
export type Vibe_tasksvibe_taskstatus = keyof typeof Vibe_tasksvibe_taskstatus;

export interface Vibe_tasksBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_tasksstatecode;
  statuscode?: Vibe_tasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_actualcost?: number;
  vibe_actualhours?: number;
  "vibe_assignedtoid@odata.bind"?: string;
  vibe_description?: string;
  vibe_duedate?: string;
  vibe_estimatedcost?: number;
  vibe_estimatedhours?: number;
  vibe_ismilestone?: boolean;
  vibe_name?: string;
  "vibe_phaseid@odata.bind"?: string;
  "vibe_predecessorid@odata.bind"?: string;
  vibe_priority?: Vibe_tasksvibe_priority;
  "vibe_projectid@odata.bind"?: string;
  vibe_startdate?: string;
  vibe_taskid: string;
  vibe_taskstatus?: Vibe_tasksvibe_taskstatus;
}

export interface Vibe_tasks extends Vibe_tasksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_assignedtoidname?: string;
  vibe_ismilestonename?: string;
  vibe_phaseidname?: string;
  vibe_predecessoridname?: string;
  vibe_priorityname?: string;
  vibe_projectidname?: string;
  vibe_taskstatusname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_assignedtoid?: object;
  _vibe_assignedtoid_value?: string;
  vibe_phaseid?: object;
  _vibe_phaseid_value?: string;
  vibe_predecessorid?: object;
  _vibe_predecessorid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}