// powerapps-project-planner-20260612/src/generated/models/Vibe_projectphasesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_projectphasesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_projectphasesstatecode = keyof typeof Vibe_projectphasesstatecode;
export const Vibe_projectphasesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_projectphasesstatuscode = keyof typeof Vibe_projectphasesstatuscode;
export const Vibe_projectphasesvibe_approvalstatus = {
  100000000: 'NOT_REQUIRED',
  100000001: 'PENDING',
  100000002: 'APPROVED',
  100000003: 'REJECTED'
} as const;
export type Vibe_projectphasesvibe_approvalstatus = keyof typeof Vibe_projectphasesvibe_approvalstatus;

export interface Vibe_projectphasesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_projectphasesstatecode;
  statuscode?: Vibe_projectphasesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_approvalstatus?: Vibe_projectphasesvibe_approvalstatus;
  vibe_enddate?: string;
  vibe_name?: string;
  vibe_order?: number;
  "vibe_projectid@odata.bind"?: string;
  vibe_projectphaseid: string;
  vibe_requiresapproval?: boolean;
  vibe_startdate?: string;
}

export interface Vibe_projectphases extends Vibe_projectphasesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_approvalstatusname?: string;
  vibe_projectidname?: string;
  vibe_requiresapprovalname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}